requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
djangorestframework==3.11.0
idna==2.10
numpy==1.19.1
pandas==1.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlalchemy==1.3.18
sqlalchemy-filters==0.12.0
sqlparse==0.3.1
urllib3==1.25.10
